PCT/requirements.txt
torch==1.8.0
torchvision==0.9.0
opencv-python
numpy
mmcv-full==1.7.0
mmpose==0.29.0
tqdm
yacs
json_tricks
tensorboardX
pycocotools   
scipy
timm
albumentations
